sandbox: rewritten to use lazy pages
