+++ packages/admin-sandbox/admin/index.tsx
@@ -8,12 +8,6 @@ runReactApp(
 		sessionToken={import.meta.env.VITE_CONTEMBER_ADMIN_SESSION_TOKEN as string}
 		project={'admin-sandbox'}
 		stage={'live'}
-		basePath={import.meta.env.BASE_URL}
-		children={
-			<Pages
-				layout={Layout}
-				children={Object.assign({}, ...Object.values(import.meta.globEager('./pages/*.tsx')))}
-			/>
-		}
+		children={<Pages layout={Layout} children={import.meta.glob('./pages/*.tsx')} />}
 	/>,
 )
