--- packages/admin-sandbox/admin/pages/SecondPage.tsx
@@ -1,7 +1,0 @@
-import { EditPage, Link } from '@contember/admin'
-
-export const Second = (
-	<EditPage  entity="UploadShowcase(unique = One)" setOnCreate="(unique = One)">
-		<Link to="index">DASHBOARD</Link>
-	</EditPage>
-)
