+++ packages/admin-sandbox/admin/pages/foo.tsx
@@ -1,0 +1,3 @@
+import { GenericPage } from '@contember/admin'
+
+export default () => <GenericPage>Hello from Foo</GenericPage>
