+++ packages/admin-sandbox/admin/pages/homepage.tsx
@@ -2,7 +2,7 @@ import { EditPage } from '@contember/admin'
 import { AddContent } from '../components/AddContent'
 import { ContentField } from '../components/ContentField'
 
-export const Homepage = (
+export default () => (
 	<EditPage entity="Homepage(unique = One)" setOnCreate="(unique = One)">
 		<ContentField field="content" />
 		<AddContent field="content" />
