+++ packages/admin-sandbox/admin/pages/lorem.tsx
@@ -1,0 +1,3 @@
+import { GenericPage } from '@contember/admin'
+
+export default () => <GenericPage>Hello from Lorem</GenericPage>
